(L1) feat: added Letizia's sol
